[update] bugfix, vector sotre

+++ src/features/build_vector_db.py
@@ -44,7 +44,7 @@ def build_vector_db(kwargs):
     logger.info(embeddings)
 
     # build vector engine
-    engine = VectorEngine(df["id"].values, embeddings)
+    engine = VectorEngine(df["id"].astype(int).values, embeddings)
 
     # search similar
     similarities, indices = engine.search(embeddings[0])
